Düello çalışır durumda. Bahis düşme ekleme yapıldı. Oyun bitince duel tablosunda score güncelleniyor

+++ src/Infrastructure/Cp/ContestPark.Infrastructure.Cp/Grains/CpGrain.cs
@@ -11,7 +11,7 @@ namespace ContestPark.Infrastructure.Cp.Grains
     {
         #region Private variables
 
-        private readonly ICpRepository _cpRepository;
+        private ICpRepository _cpRepository;
         private readonly ILogger<CpGrain> _logger;
 
         #endregion Private variables
@@ -58,17 +58,38 @@ namespace ContestPark.Infrastructure.Cp.Grains
                 return Task.FromResult(-1);
             }
 
+            if (userId.Contains("-bot"))
+                return Task.FromResult(0);
+
             int removeGold = _cpRepository.RemoveGold(userId, diminishingGold, goldProcessName);
 
             return Task.FromResult(removeGold);
         }
 
-        #endregion Methods
-
-        public override Task OnDeactivateAsync()
+        /// <summary>
+        /// Kullanıcıya altın ekler
+        /// </summary>
+        /// <param name="userId">Kullanıcı id</param>
+        /// <param name="addedChips">Eklenecek altın miktarı</param>
+        /// <param name="goldProcessName">Hangi sebepten artıyor</param>
+        /// <returns>Azaldıktan sonraki altın miktarı</returns>
+        public Task<int> AddGold(string userId, int diminishingGold, GoldProcessNames goldProcessName)
         {
-            _cpRepository.Dispose();
-            return base.OnDeactivateAsync();
+            if (string.IsNullOrEmpty(userId) || diminishingGold <= 0)
+            {
+                _logger.LogWarning($"Kullanıcı id boş veya azaltılacak bahis miktarı negatif değer. userId: {userId} gold: {diminishingGold}");
+
+                return Task.FromResult(-1);
+            }
+
+            if (userId.Contains("-bot"))
+                return Task.FromResult(0);
+
+            int addedGold = _cpRepository.AddGold(userId, diminishingGold, goldProcessName);
+
+            return Task.FromResult(addedGold);
         }
+
+        #endregion Methods
     }
 }
